+++ src/redturtle/prenotazioni/tests/test_week_table_overrides.py
@@ -129,6 +129,12 @@ class TestWeekTableOverridesContextState(unittest.TestCase):
             json.loads(self.folder_prenotazioni.week_table_overrides)[0]["week_table"],
         )
 
+        # if in range and next year +1, return table overrides
+        self.assertEqual(
+            self.view.get_week_table(date(now.year + 1, 12, 25)),
+            json.loads(self.folder_prenotazioni.week_table_overrides)[0]["week_table"],
+        )
+
         # if out of range, return base table
         self.assertEqual(
             self.view.get_week_table(date(now.year, 10, 10)),